backend/requirements.txt
flask==3.1.2
werkzeug==3.1.3
pymupdf==1.26.4
opencv-python==4.12.0.88
pytesseract==0.3.13
google-genai==1.36.0
python-dotenv==1.1.1
markdown==3.5.1
flask-cors==6.0.1
